*: refact code add SnapshotManager
