+++ src/bin/tikv-server.rs
@@ -35,14 +35,13 @@ use getopts::{Options, Matches};
 use rocksdb::{DB, Options as RocksdbOptions, BlockBasedOptions};
 use mio::tcp::TcpListener;
 use cadence::{StatsdClient, NopMetricSink, UdpMetricSink};
-use tempdir::TempDir;
 
 use tikv::storage::{Storage, Dsn, TEMP_DIR};
 use tikv::util::{self, logger, panic_hook, metric};
 use tikv::server::{DEFAULT_LISTENING_ADDR, SendCh, Server, Node, Config, bind, create_event_loop,
                    create_raft_storage};
 use tikv::server::{ServerTransport, ServerRaftStoreRouter, MockRaftStoreRouter};
-use tikv::server::{MockStoreAddrResolver, PdStoreAddrResolver};
+use tikv::server::{MockStoreAddrResolver, PdStoreAddrResolver, SnapshotManager};
 use tikv::raftstore::store;
 use tikv::pd::{new_rpc_client, RpcClient};
 
@@ -254,17 +253,14 @@ fn build_raftkv(matches: &Matches,
                 cluster_id: u64,
                 addr: String,
                 pd_client: Arc<RpcClient>)
-                -> (Storage, Arc<RwLock<ServerRaftStoreRouter>>, u64, PathBuf) {
+                -> (Storage, Arc<RwLock<ServerRaftStoreRouter>>, u64, String) {
     let trans = Arc::new(RwLock::new(ServerTransport::new(ch)));
 
     let path = get_store_path(matches, config);
-    let snap_path = Path::new(&path).join("/snapshot/").join(format!("{}/", cluster_id));
-    let snap_path_copy = snap_path.as_path().to_path_buf();
-
     let opts = get_rocksdb_option(matches, config);
 
     let engine = Arc::new(DB::open(&opts, &path).unwrap());
-    let mut cfg = Config::new(&snap_path.into_os_string().into_string().unwrap());
+    let mut cfg = Config::new();
     cfg.cluster_id = cluster_id;
 
     cfg.addr = addr.clone();
@@ -278,13 +274,19 @@ fn build_raftkv(matches: &Matches,
                                           Some(addr),
                                           |v| v.as_str().map(|s| s.to_owned()));
 
+    let snap_path = get_snap_path(matches, config);
+    cfg.store_cfg.snap_path = snap_path;
+
     let mut event_loop = store::create_event_loop(&cfg.store_cfg).unwrap();
     let mut node = Node::new(&mut event_loop, &cfg, pd_client);
     node.start(event_loop, engine.clone(), trans).unwrap();
     let raft_router = node.raft_store_router();
     let node_id = node.id();
 
-    (create_raft_storage(node, engine).unwrap(), raft_router, node_id, snap_path_copy)
+    (create_raft_storage(node, engine).unwrap(),
+     raft_router,
+     node_id,
+     cfg.store_cfg.snap_path.clone())
 }
 
 fn get_store_path(matches: &Matches, config: &toml::Value) -> String {
@@ -309,15 +311,39 @@ fn get_store_path(matches: &Matches, config: &toml::Value) -> String {
     format!("{}", absolute_path.display())
 }
 
+fn get_snap_path(matches: &Matches, config: &toml::Value) -> String {
+    let path = get_string_value("",
+                                "server.snapshot",
+                                matches,
+                                config,
+                                Some(TEMP_DIR.to_owned()),
+                                |v| v.as_str().map(|s| s.to_owned()));
+    if path == TEMP_DIR {
+        return path;
+    }
+
+    let p = Path::new(&path);
+    if p.exists() && p.is_file() {
+        panic!("{} is not a directory!", path);
+    }
+    if !p.exists() {
+        fs::create_dir_all(p).unwrap();
+    }
+    let absolute_path = p.canonicalize().unwrap();
+    format!("{}", absolute_path.display())
+}
+
 fn run_local_server(listener: TcpListener, store: Storage) {
     let mut event_loop = create_event_loop().unwrap();
     let router = Arc::new(RwLock::new(MockRaftStoreRouter));
+    let snap_manager = SnapshotManager::new(PathBuf::from("/tmp/"),
+                                            SendCh::new(event_loop.channel()));
     let mut svr = Server::new(&mut event_loop,
                               listener,
                               store,
                               router,
                               MockStoreAddrResolver,
-                              TempDir::new("snapshot").unwrap().path())
+                              Arc::new(RwLock::new(snap_manager)))
         .unwrap();
     svr.run(&mut event_loop).unwrap();
 }
@@ -346,19 +372,20 @@ fn run_raft_server(listener: TcpListener, matches: &Matches, config: &toml::Valu
     let (store, raft_router, node_id, snap_path) =
         build_raftkv(matches,
                      config,
-                     ch,
+                     ch.clone(),
                      cluster_id,
                      format!("{}", listener.local_addr().unwrap()),
                      pd_client);
 
     initial_metric(matches, config, Some(node_id));
 
+    let snap_manager = SnapshotManager::new(PathBuf::from(&snap_path), ch);
     let mut svr = Server::new(&mut event_loop,
                               listener,
                               store,
                               raft_router,
                               resolver,
-                              &snap_path)
+                              Arc::new(RwLock::new(snap_manager)))
         .unwrap();
     svr.run(&mut event_loop).unwrap();
 }
